This advancement implements UART transmission in duplex mode, that means that now we have one task for transmitting and other for receiving in the UART interface. On top of that, I have a method that initializes the interface and passes the uart handle to each location to make the use of the interface
