+++ EmbeddedDevelopment/TestingThreadsAndCommunication/.settings/org.eclipse.core.resources.prefs
@@ -1,6 +1,8 @@
 eclipse.preferences.version=1
 encoding//Debug/DataStructures/subdir_rules.mk=UTF-8
 encoding//Debug/DataStructures/subdir_vars.mk=UTF-8
+encoding//Debug/Drivers/subdir_rules.mk=UTF-8
+encoding//Debug/Drivers/subdir_vars.mk=UTF-8
 encoding//Debug/Tasks/subdir_rules.mk=UTF-8
 encoding//Debug/Tasks/subdir_vars.mk=UTF-8
 encoding//Debug/makefile=UTF-8
