+++ EmbeddedDevelopment/TestingThreadsAndCommunication/Tasks/taskDefinitions.h
@@ -1,10 +1,16 @@
+#include <ti/drivers/UART.h>
 
 //Definition of functions for LED management
-
 void *ledTask(void *arg);
 
-void *uartTask(void *arg);
+//Perform UART Write operations
+void *uartWriteTask(void *arg);
 
+//Peform UART Read operations
+void *uartReadTask(void *arg);
+
+//RF Tx task
 void *txTask(void *arg);
 
+//RF Rx Task
 void *rxTask(void *arg);
